Move new-messages.ts to /query-server
