+++ extensions/ql-vscode/src/data-extensions-editor/generate-flow-model.ts
@@ -13,7 +13,7 @@ import {
   ModeledMethodWithSignature,
 } from "./modeled-method";
 import { redactableError } from "../common/errors";
-import { QueryResultType } from "../pure/new-messages";
+import { QueryResultType } from "../query-server/new-messages";
 import { file } from "tmp-promise";
 import { writeFile } from "fs-extra";
 import { dump } from "js-yaml";
