+++ extensions/ql-vscode/src/query-server/legacy/run-queries.ts
@@ -19,7 +19,7 @@ import {
   showAndLogWarningMessage,
 } from "../../common/logging";
 import * as messages from "../legacy-messages";
-import * as newMessages from "../../pure/new-messages";
+import * as newMessages from "../new-messages";
 import * as qsClient from "./query-server-client";
 import { asError, getErrorMessage } from "../../common/helpers-pure";
 import { compileDatabaseUpgradeSequence } from "./upgrades";
