+++ extensions/ql-vscode/src/query-server/new-messages.ts
@@ -15,7 +15,7 @@
  */
 
 import { RequestType } from "vscode-jsonrpc";
-import * as shared from "../query-server/messages-shared";
+import * as shared from "./messages-shared";
 
 /**
  * Parameters to clear the cache
